data_pipelines/predictor_v1/src/requirements.txt
google-auth
google-api-python-client==2.137.0
google-cloud-bigquery[bqstorage,pandas]==3.21.0
Flask==3.0.3
gunicorn==20.1.0
google-cloud-storage==2.18.2